cleanup(devimint): move external deamons to different file
